requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0
supabase>=2.10.0
polars>=1.14.0
pandas>=2.2.0
openpyxl>=3.1.0
crewai>=0.80.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
structlog>=24.4.0
biopython>=1.83
xgboost>=2.0.0
requests>=2.31.0
